==> notebooks/Rmd/02c_seedcorr_area_stats.Rmd <==
---
title: "Seedcorr Area Stats"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    keep_md: true
    df_print: paged
---

# Seedcorr Area Stats

```{r}
library(tidyverse)
library(broom)
library(here)
```

## The paths to data

```{r setting paths}
source(here('code/R/settings_helpers.R'))

Yeo7_2011_80verts <- read_Yeo72011_template()
YeoNet_colours = define_Yeo7_colours()
pheno <- read_pheno_file() %>%  
  drop_na(DX) %>%
  drop_na(Age_match) # dropping un-matched subjects
pint_sc <- read_csv(file.path(output_base, 
                              "all_clinicalplusqa_group",
                              "postPINT",
                              'seedcorr_numvxabove_all_qa_passes.csv'))

lm_predictors <- "DX + Age_match_pt + Sex + fd_mean_match_pt + Site + SurfArea_match_pt"
```

```{r}
pint_sc %>%
  anti_join(pheno, by = c("subject", "dataset"))
```

```{r}
pheno %>%
  anti_join(pint_sc, by = c("subject", "dataset"))
```

Write a func_base and outputprefix cols into the pheno file for the file reading step

```{r}
combined_df <- pheno %>%
  inner_join(pint_sc, by = c("dataset", "subject", "session")) %>%
  gather(gt_name, num_sc_vx_gt, starts_with('gt_')) %>%
  mutate(vx_gt_thres = gt_name %>% str_replace('gt_','') %>% parse_double()) %>%
  drop_na(num_sc_vx_gt)

combined_df %>% count(subject) %>% nrow()
```

## Known issues

Looks like two people might need to be reprocessed if this works out.



```{r}
combined_df %>%
  drop_na(DX) %>%
  group_by(NETWORK, vx_gt_thres) %>%
  filter(vx_gt_thres < 0.6) %>%
  do(tidy(aov(lm(formula(paste("num_sc_vx_gt ~", lm_predictors)),.)))) %>%
  ungroup() %>% group_by(term) %>%
  mutate(p_FDR = p.adjust(p.value, method = "fdr")) %>%
  filter(term != "(Intercept)", term != "Site", p_FDR < 0.06) %>%
  arrange(NETWORK, term) 
       
```
```{r fig.height=4, fig.width=7.5}
combined_df %>%
  drop_na(DX) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres, color = as.factor(NETWORK))) +
  # geom_line(aes(group = subject), alpha = 0.1) + 
  geom_smooth() +
  facet_wrap(~Site, ncol = 4)
```

```{r fig.height=11.5, fig.width=7.5}
combined_df %>%
  drop_na(DX) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres, color = as.factor(DX))) +
  # geom_line(aes(group = subject), alpha = 0.1) + 
  geom_smooth() +
  facet_wrap(~NETWORK*Site, ncol = 4)
```

```{r fig.height=3.5, fig.width=7.5}
combined_df %>%
  drop_na(DX) %>%
  filter(NETWORK == 5) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres, color = Sex)) +
  geom_line(aes(group = subject), alpha = 0.1) + 
  geom_smooth() +
  facet_wrap(~NETWORK*Site, ncol = 4)
```
```{r fig.height = 11.5, fig.width = 7.5}
combined_df %>% 
  filter(Age > 17, Age < 51) %>%
  mutate(Age_decade = case_when(Age < 20 ~ '17-20',
                               Age >= 20 & Age < 30 ~ '20-30',
                               Age >= 30 & Age < 40 ~ '30-40',
                               Age >= 40 & Age < 50 ~ '40-50',
                               Age > 50 ~ "50 above")) %>%
  mutate(Age_coarse = case_when(Age < 30 ~ "young adult",
                                Age > 35 ~ "middle age")) %>%
  drop_na(DX, Age_coarse) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres, color = DX)) +
  #geom_line(aes(group = subject), alpha = 0.1) + 
  geom_smooth(aes(linetype = Age_coarse)) +
  facet_wrap(~NETWORK*Site, ncol = 4)  
```
```{r fig.height = 4.5, fig.width = 7.5}
combined_df %>% 
  filter(Age > 17, Age < 51) %>%
  mutate(Age_decade = case_when(Age < 20 ~ '17-20',
                               Age >= 20 & Age < 30 ~ '20-30',
                               Age >= 30 & Age < 40 ~ '30-40',
                               Age >= 40 & Age < 50 ~ '40-50',
                               Age > 50 ~ "50 above")) %>%
  mutate(Age_coarse = case_when(Age < 30 ~ "young adult",
                                Age > 35 ~ "middle age")) %>%
  drop_na(DX, Age_coarse) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres)) +
  geom_line(aes(group = subject, color = Age), alpha = 0.15) + 
  geom_smooth(aes(linetype = Age_coarse), color = "black") +
  scale_color_viridis_c(option = "D") +
  facet_grid(DX~NETWORK)  
```
```{r fig.height = 3, fig.width = 8.5}
combined_df %>% 
  filter(Age > 17, Age < 51) %>%
  mutate(Age_decade = case_when(Age < 20 ~ '17-20',
                               Age >= 20 & Age < 30 ~ '20-30',
                               Age >= 30 & Age < 40 ~ '30-40',
                               Age >= 40 & Age < 50 ~ '40-50',
                               Age > 50 ~ "50 above")) %>%
  mutate(Age_coarse = case_when(Age < 30 ~ "young adult",
                                Age > 35 ~ "middle age")) %>%
  drop_na(DX, Age_coarse) %>%
ggplot(aes(y = num_sc_vx_gt, x = vx_gt_thres)) +
  geom_line(aes(group = subject, color = Age), alpha = 0.15) + 
  geom_smooth(aes(linetype = Age_coarse), color = "black") +
  scale_color_viridis_c(option = "D") +
  facet_grid(~NETWORK)  
```

```{r}
combined_df %>% 
  filter(vx_gt_thres == "0.5") %>%
  mutate(Age_decade = case_when(Age < 20 ~ '17-20',
                               Age >= 20 & Age < 30 ~ '20-30',
                               Age >= 30 & Age < 40 ~ '30-40',
                               Age >= 40 & Age < 50 ~ '40-50',
                               Age > 50 ~ "50 above")) %>%
  mutate(Age_coarse = case_when(Age < 30 ~ "young adult",
                                Age > 35 ~ "middle age")) %>%
  drop_na(DX) %>%
ggplot(aes(y = num_sc_vx_gt, x = Age, color = DX)) +
  geom_point(alpha = 0.15) + 
  geom_smooth(method = lm) +
  facet_grid(~NETWORK) 
```

```{r}
combined_df %>% 
  filter(Age > 17, Age < 51) %>%
  mutate(Age_decade = case_when(Age < 20 ~ '17-20',
                               Age >= 20 & Age < 30 ~ '20-30',
                               Age >= 30 & Age < 40 ~ '30-40',
                               Age >= 40 & Age < 50 ~ '40-50',
                               Age > 50 ~ "50 above")) %>%
  mutate(Age_coarse = case_when(Age < 30 ~ "young adult",
                                Age > 35 ~ "middle age")) %>%
  filter(NETWORK == 5, vx_gt_thres ==0.5) %>%
  count(DX, Site, Age_coarse)
```
