==> notebooks/Rmd/05_sulcal_depth_stats.Rmd <==
---
title: "Sulcal Depth Stats"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    keep_md: true
    df_print: paged
---

# Sulcal Depth Results

```{r}
library(tidyverse)
library(broom)
```

## The paths to data

```{r setting paths}
output_base <- '../data/ciftify_fmriprep/'

Yeo7_2011_80verts <- read_csv("../templates/Yeo7_2011_80verts.csv",
                              col_types = c(
                                hemi = col_character(),
                                tvertex = col_integer(),
                                LRpairs = col_integer(),
                                roiidx = col_integer(),
                                NETWORK = col_integer(),
                                LOBE = col_character(),
                                SHORTNAME = col_character(),
                                x = col_integer(),
                                y = col_integer(),
                                z = col_integer()
                              ))

YeoNet_colours = list("VI" = "#781286",
                      "SM" = "#4682B4",
                      "DA" = "#00760E", 
                      "VA" = "#C43AFA",
                      "DM" = "#CD3E3A", 
                      "FP" = "#E69422")

pheno <- read_csv('../phenotypic/20190301_pheno_qapass.csv') 

pint_concat <- read_csv('../data/ciftify_fmriprep/postPINT1_concat_all_qa_passes.csv')
```



These functions are for reading timeseries files

```{r subcortical file reading functions}

#' read a meants file generated by PINT of ciftify_meants
#'
#' @param filepath the full path to the file
#'
#' @return a dataframe where rows are rois and colums are timepoints
read_meants_csv <- function(filepath) {
   meants <-read_csv(filepath, 
                     col_names = FALSE,
                     col_types = c(.default = col_double()))
   return(meants)
}


#' Contructs the expected output prefix from subid session and func_base
#'
#' @param subid The subject identifier
#' @param sessid The session identifier (or null)
#' @param func_base The functional file prefix
#'
#' @return an output prefix string for the filenames
construct_output_prefix <- function(subid, sessid, func_base) {
  prefix <- if_else(is.na(sessid),
                    file.path(subid, str_c(subid, '_', func_base)),
                    file.path(subid, sessid, 
                      str_c(subid, '_', sessid, '_', func_base)))
  return(prefix)
}

#' get func_base from pint summary filename
#'
#' @param subid The subject identifier
#' @param sessid The session identifier (or null)
#' @param func_base The functional file prefix
#'
#' @return an output prefix string for the filenames
get_func_base_from_pint_summary_filename <- function(filename, subject, session) {
  func_base <- if_else(is.na(session), 
                       filename %>%
                         str_replace(str_c(subject, "_"), '') %>%
                         str_replace('_desc-clean_bold_summary.csv',''),
                       filename %>%
                         str_replace(subject, '') %>%
                         str_replace(session, '') %>% 
                         str_replace('__','') %>%
                         str_replace('_desc-clean_bold_summary.csv',''))
  return(func_base)
}

```

Write a func_base and outputprefix cols into the pheno file for the file reading step

```{r}
pheno <- pheno %>%
  mutate(func_base = get_func_base_from_pint_summary_filename(filename,subject, session), 
         outputprefix = construct_output_prefix(subject, session, func_base)) 
```


## check that sulcal depth has been calculated in all participants

```{r identify-sulc-outputs}
sulc_summarys_wses <- Sys.glob("../data/ciftify_fmriprep/*/out/ciftify_PINT/sub*/ses*/*pvertexROI_morph.csv")

sulc_summarys_nses <- Sys.glob("../data/ciftify_fmriprep/*/out/ciftify_PINT/sub*/*pvertexROI_morph.csv")

sulcdf_nses <- tibble(filepath = sulc_summarys_nses) %>%
  separate(filepath,
           into = c('1','2','3','dataset','4','5','subject', 'sfilename'),
           sep = .Platform$file.sep) %>%
  select(dataset, subject, sfilename)

sulcdf_wses <- tibble(filepath = sulc_summarys_wses) %>%
  separate(filepath,
           into = c('1','2','3','dataset','4','5','subject', 'session', 'sfilename'),
           sep = .Platform$file.sep) %>%
  select(dataset, subject, session, sfilename)

sulcdf <- bind_rows(sulcdf_wses, sulcdf_nses) %>%
  mutate(filename = str_replace(sfilename, 'desc-sulc_atlas-pvertexROI_morph','desc-clean_bold_summary'))

rm(sulc_summarys_nses, sulc_summarys_wses, sulcdf_nses, sulcdf_wses)
```

```{r}
pheno %>%
  anti_join(sulcdf, by="filename")
```

## Known issues

one of the RTMSWM failed when Jerry reprossed them all..so now it is missing surface..urg..will exclude for now

```{r}
pheno <- pheno %>% filter(subject != "sub-CMHWM040")
```


```{r}
read_sulc_depth <- function(dataset, outputprefix) {
  pvertex_path <- file.path(output_base, dataset, 'out','ciftify_PINT',
                                  str_c(outputprefix, '_desc-sulc_atlas-pvertexROI_morph.csv'))
  pvertex_sulc <- read_meants_csv(pvertex_path) %>% rename(pvertex_sulc = X1)
  tvertex_path <- str_replace(pvertex_path, 'pvertexROI','tvertexROI')
  tvertex_sulc <- read_meants_csv(tvertex_path) %>% rename(tvertex_sulc = X1)
  results <- bind_cols(roiidx = 1:80, pvertex_sulc, tvertex_sulc)
}

all_sulcs <- pheno %>%
  select(dataset, subject, fd_mean_pt, fd_perc, DX, Age, Age_pt, Sex, Site, Scanner, outputprefix, filename) %>%
  mutate(sulcs = map2(dataset, outputprefix, ~read_sulc_depth(.x,.y))) %>%
  unnest()

```

```{r}
all_sulcs %>%
  drop_na(DX) %>%
  group_by(roiidx, Site) %>%
  do(tidy(lm(tvertex_sulc ~ DX + Sex + fd_mean_pt + poly(Age_pt,2),.))) %>%
  ungroup() %>% group_by(term) %>%
  mutate(p_FDR = p.adjust(p.value, method = "fdr")) %>%
  inner_join(Yeo7_2011_80verts, by = "roiidx") %>%
  filter(term != "(Intercept)", p_FDR < 0.1) %>%
  arrange(p.value)
       
```
```{r fig.height=11.5, fig.width=7.5}
all_sulcs %>%
  drop_na(DX) %>%
  inner_join(Yeo7_2011_80verts, by = "roiidx") %>%
  ggplot(aes(x = tvertex_sulc, fill = DX)) + geom_density(alpha = 0.5) + facet_grid(Site~LOBE) 
```

```{r}
all_sulcs %>%
  drop_na(DX) %>%
  filter(roiidx == 68) %>%
  ggplot(aes(x = tvertex_sulc, y = fd_perc, color = Site)) + geom_point()
```


```{r}
concat_plus_sulc <- all_sulcs %>% 
  mutate(subid = str_replace(filename, '_summary.csv','')) %>%
  inner_join(pint_concat, by = c("subid", "roiidx"))
```

```{r}
dist_plus_sulc_by_net <- concat_plus_sulc %>%
  mutate(diff_sulc = pvertex_sulc - tvertex_sulc) %>%
  group_by(roiidx, Site) %>%
  summarise(mean_tsulc = mean(tvertex_sulc),
    mean_psulc = mean(pvertex_sulc),
    sd_tsulc = sd(tvertex_sulc),
            sd_psulc = sd(pvertex_sulc),
            sd_diffsulc = sd(diff_sulc),
            mean_diffsulc = mean(diff_sulc),
            mean_stddistance = mean(std_distance),
            sd_stddistance = sd(std_distance)) %>%
  inner_join(Yeo7_2011_80verts, by = "roiidx") %>%
  mutate(network = str_sub(SHORTNAME, 1,2))
```

```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_stddistance, y = sd_tsulc)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_wrap(~Site, ncol = 4)
```
```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_stddistance, y = sd_tsulc)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_grid(Site~network)
```
```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_stddistance, y = sd_tsulc)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_grid(Site~LOBE)
```
```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_stddistance, y = sd_psulc)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_wrap(~Site, ncol = 4)
```

```{r}
dist_plus_sulc_by_net_dx <- concat_plus_sulc %>%
  mutate(diff_sulc = pvertex_sulc - tvertex_sulc) %>%
  group_by(roiidx, Site, DX) %>%
  summarise(sd_tsulc = sd(tvertex_sulc),
            sd_psulc = sd(pvertex_sulc),
            sd_diffsulc = sd(diff_sulc),
            mean_diffsulc = mean(diff_sulc),
            mean_stddistance = mean(std_distance),
            sd_stddistance = sd(std_distance)) %>%
  inner_join(Yeo7_2011_80verts, by = "roiidx") %>%
  mutate(network = str_sub(SHORTNAME, 1,2))
```

```{r}
dist_plus_sulc_by_net_dx %>%
  ggplot(aes(x = mean_stddistance, y = sd_tsulc, color = DX)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_grid(~Site)
```

```{r}
dist_plus_sulc_by_net_dx %>%
  ggplot(aes(x = sd_stddistance, y = sd_tsulc, color = DX)) +
  geom_point() +
  geom_smooth(method = "lm") +
  facet_wrap(~Site, ncol = 4)
```

```{r}
dist_plus_sulc_by_net %>%
  ungroup() %>%
  select(mean_tsulc:sd_stddistance) %>%
  cor()
```

```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_diffsulc, y = mean_tsulc, color = LOBE)) + geom_point()
```

```{r}
dist_plus_sulc_by_net %>%
  ggplot(aes(x = mean_psulc, y = mean_tsulc, color = LOBE)) + geom_point() +
  geom_abline(slope = 1, intercept = 0)
```
